+++ src/utils/queries.js
@@ -34,6 +34,18 @@ const eventFields = `
   }
 `;
 
+// export const GET_LATEST_EVENTS_BY_ID = (id) => {
+//   debugger;
+//   return gql`
+//   query {
+//     miningEvents(first: 10, orderBy: timestamp, orderDirection: desc, where: { requestIds_contains: [`+ id + `] }) {
+//       ${eventFields}
+//     }
+//   }
+// `;
+// }
+
+
 export const GET_LATEST_EVENTS = gql`
   query {
     miningEvents(first: 6, orderBy: timestamp, orderDirection: desc) {
